renamed algebras into algebra
