+++ public/models/os-project.js
@@ -2,7 +2,7 @@ import set from "can-set";
 import DefineMap from "can-define/map/";
 import DefineList from "can-define/list/";
 import superModel from '../lib/super-model';
-import algebra from './algebras';
+import algebra from './algebra';
 import feathersClient from './feathers-client';
 
 var OSProject = DefineMap.extend("OSProject", { seal: false }, {
