+++ public/models/user.js
@@ -3,7 +3,7 @@ import DefineMap from 'can-define/map/';
 import DefineList from 'can-define/list/list';
 import feathersClient from './feathers-client';
 import superModel from '../lib/super-model';
-import algebra from './algebras';
+import algebra from './algebra';
 
 var User = DefineMap.extend("User", {
 	_id: "string",
